+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 8.52.0 <span class="changelog-date">(2024-12-22)</span>
+
 ## 8.51.0 <span class="changelog-date">(2024-12-15)</span>
 
 * Added a wishlist refresh button in Settings to help with wishlist development. Note that GitHub can take upwards of 10min to actually reflect your changes. Refreshing won't speed up GitHub.